+++ Extractor/Models/UserEngagement.cs
@@ -5,7 +5,7 @@ namespace Extractor.Models
     public class UserEngagement
     {
         public string Username { get; set; }
-        public int TotalComments { get; set; }
+        public long TotalComments { get; set; }
         public Dictionary<string, float> CommunityProportions { get; set; }
     }
 }
